requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
google-genai==0.3.0
requests==2.31.0
Pillow==10.1.0
opencv-python==4.8.1.78
